% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Koboconnect.R
\name{get_kobo_id}
\alias{get_kobo_id}
\title{Retrieve kobotools id from Renviron which was set using `set_kobo_id_pass`}
\usage{
get_kobo_id()
}
\value{
Returns the original id as a string
}
\description{
`get_kobo_id` retrieves your kobotools id from Renviron, which was set using `set_kobo_id_pass`.
}
\details{
The function does not take in input. It simply returns the kobotools id saved in the Renviron.
}
\examples{
get_kobo_id()

}
\seealso{
[set_kobo_id_pass()]
}
