% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Koboconnect.R
\name{get_kobo_pass}
\alias{get_kobo_pass}
\title{Retrieve kobotools password from Renviron which was set using `set_kobo_id_pass`}
\usage{
get_kobo_pass()
}
\value{
Returns the original password as a string
}
\description{
`get_kobo_pass` retrieves your kobotools password from Renviron, which was set using `set_kobo_id_pass`.
}
\details{
The function does not take in input. It simply returns the kobotools password saved in the Renviron.
}
\examples{
get_kobo_pass()

}
\seealso{
[set_kobo_id_pass()]
}
