% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Koboconnect.R
\name{get_kobo_token}
\alias{get_kobo_token}
\title{Extract data from kobotools}
\usage{
get_kobo_token(url = "kobo.humanitarianresponse.info", uname = "", pwd = "")
}
\arguments{
\item{url}{The `[url]` of kobotools. Default is "kobo.humanitarianresponse.info".}

\item{uname}{is username of your kobotool account}

\item{pwd}{is the password of the account}
}
\value{
The function returns the token associated with your id and password in the given url.
}
\description{
`get_kobo_token` is a wrapper for kobotools API `https://"[url]"/token/?format=json`
}
\details{
The function returns the API token.
}
