% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Koboconnect.R
\name{kobotools_api}
\alias{kobotools_api}
\title{Check kobotools API and retrieve overall info about the projects/assets}
\usage{
kobotools_api(
  url = "kobo.humanitarianresponse.info",
  simplified = F,
  kobo.token = NULL
)
}
\arguments{
\item{url}{The `[kpi-url]` of kobotools. Default is "kobo.humanitarianresponse.info"}

\item{simplified}{A logical value, default is false}

\item{kobo.token}{a string (the token value)}
}
\value{
The function returns the asset details from the API, inform of a data frame or json.
}
\description{
`kobotools_api` is a wrapper for kobotools API `https://[kpi-url]/api/v2/assets.json`
}
\details{
The function takes two variables. First one is `url` which is the `[kpi-url]`. For most users it will be "kobo.humanitarianresponse.info" or
"kf.kobotoolbox.org". Former is the default. THe second parameter is `simplified` which takes a logical value. If set to true,
the function will return selected values from the parsed data and return a data frame. WHen set to false, a json will be reurned with
all the details. Token needs to be set before the function can be used.
}
