% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Koboconnect.R
\name{kobotools_data_list_kc}
\alias{kobotools_data_list_kc}
\title{Lists available data in your kobotool}
\usage{
kobotools_data_list_kc(
  url = "kc.humanitarianresponse.info",
  user.id = "id",
  pass = "password"
)
}
\arguments{
\item{url}{The `[kc-url]` of kobotools. Default is "kc.humanitarianresponse.info"}

\item{user.id}{takes the user id in string}

\item{pass}{takes the password as string}
}
\value{
The function returns the list of assets/projects in kobotools, in form of a json.
}
\description{
`kobotools_data_list_kc` is a wrapper for kobotools API `https://kc.humanitarianresponse.info/api/v1/data`
}
\details{
The function takes the `kc` url as the input. It retrieves the user id and password set using `set_kobo_id_pass`.
This function uses the old api (v1) of kobotools. Similar version of v2 and `kpi` was not found in the documentation yet.
An alternative way to use the `kpi` and v2 is to use  [kobotools_data_list_kpi()]. The utility of both these function is to
find out the link to the submitted data files. In this function the returned `id` from the asset list is particularly
important which will be used in [kobotools_kc_download()] to download csv files.
}
\examples{
/notrun
kobotools_data_list_kc("kc.humanitarianresponse.info", user.id="id", pass="password")

}
\seealso{
[kobotools_kc_download()]
}
