% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Koboconnect.R
\name{kobotools_kc_download}
\alias{kobotools_kc_download}
\title{Download data from kobotools}
\usage{
kobotools_kc_download(
  assetid = NULL,
  url = "kc.humanitarianresponse.info",
  user.id = get_kobo_id(),
  pass = get_kobo_pass()
)
}
\arguments{
\item{assetid}{is the asset id of the asset for which the data is to be downloaded. The id can be found by running [kobotools_data_list_kc()]}

\item{url}{The `[kc-url]` of kobotools. Default is "kc.humanitarianresponse.info"}

\item{user.id}{takes the user id in string}

\item{pass}{takes the password as string}
}
\value{
The function returns the data
}
\description{
`kobotools_kc_download` is a wrapper for kobotools API `https://kc.humanitarianresponse.info/api/v1/data/[assetid]`
}
\details{
The function takes the `kc` url as one of the inputs. And assetid as another. Both are strings. The asset id is found by running
the [kobotools_data_list_kc()] function.
}
\examples{
/notrun
kobotools_kc_download(assetid="12345")

}
\seealso{
[kobotools_data_list_kc()]
}
