% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Koboconnect.R
\name{kobotools_kpi_data}
\alias{kobotools_kpi_data}
\title{Extract data from kobotools}
\usage{
kobotools_kpi_data(
  assetid,
  url = "kobo.humanitarianresponse.info",
  uname = "",
  pwd = ""
)
}
\arguments{
\item{assetid}{is the asset id of the asset for which the data is to be downloaded. The id can be found by running [kobotools_data_list_kc()]}

\item{url}{The `[kpi-url]` of kobotools. Default is "kobo.humanitarianresponse.info"}

\item{uname}{is username of your kobotool account}

\item{pwd}{is the password of the account}
}
\value{
The function returns the data
}
\description{
`kobotools_kpi_data` is a wrapper for kobotools API `https://[URL]/api/v2/assets/{assetid}/data/`
}
\details{
The function takes the url as one of the inputs. And asset id as another. Both are strings. The asset id is found by running
the [kobotools_api()] function.Other parameters are username and password.
}
