% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Koboconnect.R
\name{set_kobo_id_pass}
\alias{set_kobo_id_pass}
\title{Set userid and password}
\usage{
set_kobo_id_pass(id = NULL, password = NULL, samplekey = "abracadabra")
}
\arguments{
\item{id}{The Kobotoools id as string}

\item{password}{The Kobotools password as string}

\item{samplekey}{A string that you want to set as key}
}
\value{
The function encrypts the id and password provided and stores in Renviron.
}
\description{
`set_kobo_id_pass` encrypts your kobotools id and password in Renviron. It can later be retrieved by using `get_kobo_id` and `get_kobo_pass`.
}
\details{
The function takes three parameters. The first is the `id` which must be a string/character wrapped in double quotes.
The second one is the `password` which must also be a string.
The third is the `samplekey`. The default is set to "abracadabra". One may change it.
}
\examples{
set_kobo_id_pass("myuserid","mypassword")
set_kobo_token("myuserid","mypassword", "mykey")

}
\seealso{
[get_kobo_id()], [get_kobo_pass()]
}
