% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Koboconnect.R
\name{set_kobo_token}
\alias{set_kobo_token}
\title{Set encrypted token in Renviron}
\usage{
set_kobo_token(token = NULL, samplekey = "abracadabra")
}
\arguments{
\item{token}{The Kobotoools API token as string}

\item{samplekey}{A string that you want to set as key}
}
\value{
The function encrypts the token using the key provided and stores in Renviron.
}
\description{
`set_kobo_token` encrypts your token and stores in Renviron. It can later be retrieved by using `get_kobo_token`.
}
\details{
The function takes two parameters. The first is the `token` which must be a string/character wrapped in double quotes.
The second is the `samplekey`. The default is set to "abracadabra". One may change it.
}
\examples{
set_kobo_token("d7a1faf575e047d4ebb1ed")
set_kobo_token("mykobotoolstoken", "mykey")

}
\seealso{
[get_kobo_token()]
}
