% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Koboconnect.R
\name{get_kobo_token}
\alias{get_kobo_token}
\title{Know your API token or check}
\usage{
get_kobo_token(
  url = "eu.kobotoolbox.org",
  uname = "",
  pwd = "",
  encoding = "UTF-8"
)
}
\arguments{
\item{url}{The `[url]` of kobotoolbox. Default is "eu.kobotoolbox.org".}

\item{uname}{is username of your kobotoolbox account}

\item{pwd}{is the password of the account}

\item{encoding}{is the encoding to be used. Default is "UTF-8".}
}
\value{
The function returns the token associated with your id and password in the given url.
}
\description{
`get_kobo_token` is a wrapper for kobotoolbox API `https://"[url]"/token/?format=json`
}
\details{
The function returns the API token.
}
