% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Koboconnect.R
\name{get_kobo_token}
\alias{get_kobo_token}
\title{Retrieve encrypted token from Renviron}
\usage{
get_kobo_token()
}
\value{
Returns the original token as a string
}
\description{
`get_kobo_token` retrieves your kobotools token from Renviron, which was set using `set_kobo_token`.
}
\details{
The function does not take in input. It simply returns the kobotools token saved in the Renviron.
}
\examples{
get_kobo_token()

}
\seealso{
[set_kobo_token()]
}
