% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Koboconnect.R
\name{kobo_exports}
\alias{kobo_exports}
\title{See list of exports created}
\usage{
kobo_exports(
  url = "eu.kobotoolbox.org",
  uname = "",
  pwd = "",
  encoding = "UTF-8"
)
}
\arguments{
\item{url}{The `[url]` of kobotoolbox. Default is "eu.kobotoolbox.org".}

\item{uname}{is username of your kobotoolbox account}

\item{pwd}{is the password of the account}

\item{encoding}{is the encoding to be used. Default is "UTF-8".}
}
\value{
The function returns a list of exports available for the account id and password entered.
}
\description{
`kobo_exports` is a wrapper for kobotoolbox API `https://[url]/exports/`
}
\details{
The function returns the export views.
}
