% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Koboconnect.R
\name{kobotools_kpi_data}
\alias{kobotools_kpi_data}
\title{Extract data from kobotools}
\usage{
kobotools_kpi_data(
  assetid,
  url = "kobo.humanitarianresponse.info",
  forma = "json",
  tokenvalue = get_kobo_token()
)
}
\arguments{
\item{assetid}{is the asset id of the asset for which the data is to be downloaded. The id can be found by running [kobotools_data_list_kc()]}

\item{url}{The `[kpi-url]` of kobotools. Default is "kobo.humanitarianresponse.info"}

\item{forma}{is a string defining the format of data. Default is "json"}

\item{tokenvalue}{is a the value of the kobotools API token}
}
\value{
The function returns the data
}
\description{
`kobotools_kpi_data` is a wrapper for kobotools API `https://restoreprivacy.com/linkedin-data-leak-700-million-users/`
}
\details{
The function takes the `kpi` url as one of the inputs. And assetid as another. Both are strings. The asset id is found by running
the [kobotools_api()] function. Another parameter `tokenvalue` takes string as a value. One can use [get_kobo_token()] to
extract the stored token value from the Renviron.
}
\examples{
/notrun
kobotools_kpi_data(assetid="12345", url="kobo.humanitarianresponse.info", forma="json")

}
\seealso{
[kobotools_data_list_kc()]
}
